feat(base): move specific classes to different assembly
