+++ src/Christofel.BaseLib.Implementations/Configuration/BotOptions.cs
