+++ src/Christofel.BaseLib.Implementations/Extensions/IRoleExtensions.cs
