+++ src/Christofel.BaseLib.Implementations/Extensions/IUserExtensions.cs
