+++ src/Christofel.BaseLib.Implementations/Extensions/PermissionResolverExtensions.cs
