+++ src/Christofel.HelloWorld/Christofel.HelloWorld.csproj
@@ -25,6 +25,7 @@
           <Private>false</Private>
           <ExcludeAssets>runtime</ExcludeAssets>
       </ProjectReference>
+      <ProjectReference Include="..\Christofel.BaseLib.Implementations\Christofel.BaseLib.Implementations.csproj" />
       <ProjectReference Include="..\Christofel.BaseLib\Christofel.BaseLib.csproj">
           <Private>false</Private>
           <ExcludeAssets>runtime</ExcludeAssets>
